AI-Corporate-Agent-v0/app/requirements.txt
streamlit==1.36.0
python-docx==1.1.2
langchain-text-splitters==0.2.2
PyPDF2==3.0.1
numpy==1.26.4
httpx==0.27.0
python-dotenv==1.0.1
openai==1.40.2 